move const structs to ledcontrol
