+++ main/CMakeLists.txt
@@ -1,4 +1,4 @@
 idf_component_register(
-    SRCS "consts.c" "main.c" "ledcontrol.c" "wifi.c" "graph_client.c"
+    SRCS "main.c" "ledcontrol.c" "wifi.c" "graph_client.c"
     INCLUDE_DIRS "include"
 )
